+++ packages/zent-select/src/Popup.js
@@ -3,39 +3,50 @@
  */
 
 import React, { Component, PropTypes } from 'react';
-import assign from 'zent-utils/lodash/assign';
 import Search from './components/Search';
 import Option from './components/Option';
 import { KEY_EN, KEY_UP, KEY_DOWN } from './constants';
-
-const isArray = function (o) {
-  return Object.prototype.toString.apply(o) === '[object Array]';
-};
+import isArray from 'zent-utils/lodash/isArray';
 
 class Popup extends Component {
 
   constructor(props) {
     super(props);
 
-    this.state = assign({}, props);
+    this.state = {
+      data: [],
+      keyCode: '',
+      keyword: ''
+    };
     this.sourceData = props.data;
     this.searchFilterHandler = this.searchFilterHandler.bind(this);
     this.optionChangedHandler = this.optionChangedHandler.bind(this);
   }
 
   componentWillReceiveProps(nextProps) {
-    let nextState = assign({}, nextProps);
-    nextState.keyword = this.state.keyword;
-    nextState.value = this.props.extraFilter ? nextProps.value : this.state.value;
     this.sourceData = nextProps.data;
     if (nextProps.keyCode === KEY_EN && this.state.keyCode === nextProps.keyCode) {
       return;
     }
     this.updateCurrentId(nextProps.keyCode, nextProps.keyword);
-    this.setState(nextState);
+    if (nextProps.keyword === null) {
+      this.setState({
+        keyCode: nextProps.keyCode,
+        data: nextProps.data
+      });
+    } else {
+      this.setState({
+        keyCode: nextProps.keyCode,
+        data: nextProps.data,
+        keyword: nextProps.keyword
+      });
+    }
   }
 
   optionChangedHandler(ev, cid) {
+    this.setState({
+      keyword: ''
+    });
     this.props.onBlur();
     this.props.onChange(ev, this.props.data.filter(item => item.cid === cid)[0]);
   }
@@ -105,25 +116,35 @@ class Popup extends Component {
     } = this.props;
 
     let {
-      value,
-      keyword
+      keyword,
+      data
     } = this.state;
 
-    let { data } = this.state;
-    let filterData = data.filter(item => extraFilter && !value || !extraFilter && !keyword || !filter || filter(item, extraFilter ? `${value}` : `${keyword}`));
+    let filterData = data.filter(item => {
+      return !keyword || !filter || filter(item, `${keyword}`);
+    });
     let showEmpty = data.length === 0 || filterData.length === 0;
 
     this.itemIds = filterData.map(item => item.cid);
 
     return (
       <div tabIndex="0" className={`${prefixCls}-popup`} onFocus={onFocus} onBlur={onBlur}>
-        {!extraFilter && filter && <Search prefixCls={prefixCls} placeholder={searchPlaceholder} onChange={this.searchFilterHandler} />}
-        {filterData.map((item) => {
+        {
+          !extraFilter && filter ? (
+            <Search
+              keyword={keyword}
+              prefixCls={prefixCls}
+              placeholder={searchPlaceholder}
+              onChange={this.searchFilterHandler}
+            />
+          ) : ''
+        }
+        {filterData.map((item, index) => {
           let currentCls = typeof this.currentId !== 'undefined' && item.cid === this.currentId ? 'current' : '';
           let activeCls = selectedItems.filter(o => o.cid === item.cid).length > 0 || item.cid === cid ? 'active' : '';
           return (
             <Option
-              key={item.cid}
+              key={index}
               className={`${prefixCls}-option ${activeCls} ${currentCls}`}
               {...item}
               onClick={this.optionChangedHandler}
@@ -142,7 +163,6 @@ class Popup extends Component {
 
 Popup.propTypes = {
   cid: PropTypes.string,
-  value: PropTypes.any,
   keyword: PropTypes.any,
   selectedItems: PropTypes.array,
   searchPlaceholder: PropTypes.string,
@@ -156,8 +176,7 @@ Popup.propTypes = {
 };
 
 Popup.defaultProps = {
-  cid: 0,
-  value: '',
+  cid: -1,
   keyword: '',
   selectedItems: [],
   emptyText: '',